Moved files into sub-project folders
