+++ legion_core/src/iterator.rs
