+++ legion_core/src/storage.rs
