+++ legion_serialization/src/ser.rs
