+++ legion_systems/src/resource.rs
